LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-once.window-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '}'
CONSOLE MESSAGE: SyntaxError: Unexpected token '}'

FAIL Invalid uncompiled raw handlers should only be compiled once assert_equals: expected 1 but got 2
